+++ crates/filter-parser2/src/lib.rs
@@ -62,18 +62,12 @@
 //! - IN context
 //! - TO context
 
-use std::borrow::Cow;
 use std::collections::BTreeMap;
 
-use itertools::Itertools;
-use nom::branch::alt;
-use nom::bytes::complete::{escaped, tag, take_while, take_while1};
-use nom::character::complete::{anychar, multispace0, one_of};
-use nom::combinator::cut;
-use nom::number::complete::recognize_float;
-use nom::sequence::{preceded, terminated};
-use nom::{IResult, Input, Parser};
-use wip::WipResultExt;
+mod span;
+mod token;
+
+pub use span::{Span, SpanView, SpanViewError};
 
 type SourceHandle = u16;
 
@@ -93,25 +87,6 @@ pub struct FilterSource {
     source: String,
     previous_link: Link,
     next_down: u16,
-}
-
-/// A reference inside of a source.
-///
-/// Consists of a handle to the source and the span inside the source.
-/// The span must be smaller than 4GB.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Span {
-    source: SourceHandle,
-    start: u32,
-    end: u32,
-}
-
-impl Span {
-    /// Build a span from a source and its handle, spanning the entire source  
-    fn from_entire_source(source_handle: SourceHandle, source: &FilterSource) -> Self {
-        let end = source.source.len().try_into().unwrap_wip();
-        Self { source: source_handle, start: 0, end }
-    }
 }
 
 /// A filter contains sources, and instructions obtained by parsing the sources.
@@ -318,327 +293,6 @@ enum ParsingState<'a> {
     To { left_operand: SpanView<'a> },
 }
 
-#[derive(Debug, Clone, Copy)]
-struct Token<'a> {
-    span: SpanView<'a>,
-    kind: TokenKind,
-}
-
-struct ParseOutput<'a> {
-    parsed_token: Token<'a>,
-    remaining_input: SpanView<'a>,
-}
-
-impl<'a> Token<'a> {
-    /// Parses the first token of an input span, returning the parsed token and the remaining input.
-    pub fn parse_next(input: SpanView<'a>) -> Result<ParseOutput<'a>, ParseTokenError> {
-        let input = ignore_whitespace(input);
-        let (remaining_input, parsed_token) = parse_token().parse(input)?;
-        Ok(ParseOutput { parsed_token, remaining_input })
-    }
-
-    /// Iterates over all parsed tokens from input until reaching EOF.
-    pub fn iter_tokens(
-        input: SpanView<'a>,
-    ) -> impl Iterator<Item = Result<Token<'a>, ParseTokenError>> + 'a {
-        let input = std::cell::Cell::new(input);
-        std::iter::from_fn(move || {
-            let token = (|| {
-                let ParseOutput { parsed_token, remaining_input } = Self::parse_next(input.get())?;
-                input.set(remaining_input);
-                Ok(parsed_token)
-            })();
-            Some(token)
-        })
-        .take_while_inclusive(|token| {
-            let Ok(token) = token else {
-                return false;
-            };
-            !matches!(token.kind, TokenKind::Eof)
-        })
-    }
-}
-
-impl<'a> From<nom::Err<nom::error::Error<SpanView<'a>>>> for ParseTokenError {
-    fn from(value: nom::Err<nom::error::Error<SpanView<'a>>>) -> Self {
-        match value {
-            nom::Err::Incomplete(needed) => ParseTokenError::TruncatedInput(needed),
-            nom::Err::Error(err) | nom::Err::Failure(err) => {
-                wip::fixme!("proper structured error handling");
-                ParseTokenError::Error(format!("{err:?}"))
-            }
-        }
-    }
-}
-
-fn ignore_whitespace(input: SpanView<'_>) -> SpanView<'_> {
-    match nom::character::complete::multispace0::<_, ()>(input) {
-        Ok((rest, _)) => rest,
-        Err(_) => input,
-    }
-}
-
-fn is_value_component(c: char) -> bool {
-    c.is_alphanumeric() || ['_', '-', '.'].contains(&c)
-}
-
-fn parse_token<'a>(
-) -> impl Parser<SpanView<'a>, Output = Token<'a>, Error = nom::error::Error<SpanView<'a>>> {
-    parse_all_syntax()
-        .or(parse_keywords())
-        .or(parse_reserved_fields())
-        .or(parse_eof)
-        .or(parse_value)
-        .or(parse_illegal)
-}
-
-fn parse_illegal<'a>(input: SpanView<'a>) -> IResult<SpanView<'a>, Token<'a>> {
-    let (remaining, illegal) = input.take_split(1);
-    Ok((remaining, Token { span: illegal, kind: TokenKind::IllegalCharacter }))
-}
-
-fn parse_keywords<'a>(
-) -> impl Parser<SpanView<'a>, Output = Token<'a>, Error = nom::error::Error<SpanView<'a>>> {
-    alt((
-        parse_keyword("OR", TokenKind::Or),
-        parse_keyword("AND", TokenKind::And),
-        parse_keyword("NOT", TokenKind::Not),
-        parse_keyword("NULL", TokenKind::Null),
-        parse_keyword("EXISTS", TokenKind::Exists),
-        parse_keyword("IS", TokenKind::Is),
-        parse_keyword("TO", TokenKind::To),
-        parse_keyword("IN", TokenKind::In),
-    ))
-}
-
-fn parse_all_syntax<'a>(
-) -> impl Parser<SpanView<'a>, Output = Token<'a>, Error = nom::error::Error<SpanView<'a>>> {
-    alt((
-        // start with two-characters syntax to ensure we don't parse them as separate tokens
-        parse_syntax("!=", TokenKind::Different),
-        parse_syntax(">=", TokenKind::GreaterOrEqual),
-        parse_syntax("<=", TokenKind::LowerOrEqual),
-        // single-char tokens
-        parse_syntax("(", TokenKind::LeftParens),
-        parse_syntax(")", TokenKind::RightParens),
-        parse_syntax("[", TokenKind::LeftSquareBracket),
-        parse_syntax("]", TokenKind::RightSquareBracket),
-        parse_syntax(",", TokenKind::Comma),
-        parse_syntax("=", TokenKind::Equal),
-        parse_syntax(">", TokenKind::GreaterThan),
-        parse_syntax("<", TokenKind::LowerThan),
-    ))
-}
-
-fn parse_reserved_fields<'a>(
-) -> impl Parser<SpanView<'a>, Output = Token<'a>, Error = nom::error::Error<SpanView<'a>>> {
-    alt((
-        parse_reserved_field("_geoRadius", TokenKind::GeoRadius),
-        parse_reserved_field("_geoBoundingBox", TokenKind::GeoBoundingBox),
-        parse_reserved_field("_geoPolygon", TokenKind::GeoPolygon),
-        parse_reserved_field("_vectors", TokenKind::Vectors),
-        parse_reserved_field("_foreign", TokenKind::Foreign),
-    ))
-}
-
-/// Parse the specified keyword, which should NOT be followed by a value component
-fn parse_keyword<'a>(
-    keyword: &str,
-    keyword_kind: TokenKind,
-) -> impl Fn(SpanView<'_>) -> IResult<SpanView<'_>, Token<'_>> + '_ {
-    move |input| {
-        let (rest, parsed) = nom::bytes::complete::tag(keyword)(input)?;
-        if let Some(next) = rest.iter_elements().next() {
-            if is_value_component(next) {
-                return Err(nom::Err::Error(nom::error::Error::new(
-                    input,
-                    nom::error::ErrorKind::Tag,
-                )));
-            }
-        }
-        Ok((rest, Token { span: parsed, kind: keyword_kind }))
-    }
-}
-
-/// Parse the specified syntax character, which can be followed by anything
-fn parse_syntax<'a>(
-    c: &str,
-    kind: TokenKind,
-) -> impl Fn(SpanView<'_>) -> IResult<SpanView<'_>, Token<'_>> + '_ {
-    move |input| {
-        let (rest, parsed) = nom::bytes::complete::tag(c)(input)?;
-        Ok((rest, Token { span: parsed, kind }))
-    }
-}
-
-/// Parse the specified reserved field name, which can be NOT be followed by a value component except dot
-fn parse_reserved_field<'a>(
-    field: &str,
-    kind: TokenKind,
-) -> impl Fn(SpanView<'_>) -> IResult<SpanView<'_>, Token<'_>> + '_ {
-    wip::fixme!("parse_float");
-    move |input| {
-        let (rest, parsed) = nom::bytes::complete::tag(field)(input)?;
-        if let Some(next) = rest.iter_elements().next() {
-            if next != '.' && is_value_component(next) {
-                return Err(nom::Err::Error(nom::error::Error::new(
-                    input,
-                    nom::error::ErrorKind::Tag,
-                )));
-            }
-        }
-        Ok((rest, Token { span: parsed, kind }))
-    }
-}
-
-fn parse_eof<'a>(input: SpanView<'a>) -> IResult<SpanView<'a>, Token<'a>> {
-    let (rest, parsed) = nom::combinator::eof(input)?;
-    Ok((rest, Token { span: parsed, kind: TokenKind::Eof }))
-}
-
-fn parse_value<'a>(input: SpanView<'a>) -> IResult<SpanView<'a>, Token<'a>> {
-    let value = nom::combinator::map(take_while1(is_value_component), |span| Token {
-        span,
-        kind: TokenKind::Value,
-    });
-
-    let float_value =
-        nom::combinator::map(recognize_float, |span| Token { span, kind: TokenKind::FloatValue });
-
-    let (rest, parsed) =
-        terminated(alt((float_value, value, parse_quoted_value)), multispace0).parse(input)?;
-    Ok((rest, parsed))
-}
-
-/// Inspired by <https://github.com/rust-bakery/nom/blob/main/examples/string.rs>, *sans* the output as String
-/// because we're solely interested in emitting tokens at this stage
-fn parse_quoted_value<'a>(input: SpanView<'a>) -> IResult<SpanView<'a>, Token<'a>> {
-    let not_double_quote_slash = nom::bytes::streaming::is_not("\"\\");
-    let not_single_quote_slash = nom::bytes::streaming::is_not("'\\");
-    let parse_escaped_double_quote = alt((tag("\\\""), tag("\\\\")));
-    let parse_escaped_single_quote = alt((tag("\\'"), tag("\\\\")));
-
-    let parse_double_quote_fragment = alt((
-        nom::combinator::map(not_double_quote_slash, |_| TokenKind::Value),
-        nom::combinator::map(parse_escaped_double_quote, |_| TokenKind::Value),
-        nom::combinator::map(tag("\\"), |_| TokenKind::IllegalDoubleQuoted),
-    ));
-
-    let parse_single_quote_fragment = alt((
-        nom::combinator::map(not_single_quote_slash, |_| TokenKind::Value),
-        nom::combinator::map(parse_escaped_single_quote, |_| TokenKind::Value),
-        nom::combinator::map(tag("\\"), |_| TokenKind::IllegalSingleQuoted),
-    ));
-
-    let parse_single_quoted_value = nom::multi::fold(
-        0..,
-        parse_single_quote_fragment,
-        || TokenKind::Value,
-        |previous, kind| match (previous, kind) {
-            (TokenKind::IllegalSingleQuoted, _) => TokenKind::IllegalSingleQuoted,
-            (_, TokenKind::IllegalSingleQuoted) => TokenKind::IllegalSingleQuoted,
-            (_, _) => kind,
-        },
-    );
-
-    let parse_double_quoted_value = nom::multi::fold(
-        0..,
-        parse_double_quote_fragment,
-        || TokenKind::Value,
-        |previous, kind: TokenKind| match (previous, kind) {
-            (TokenKind::IllegalDoubleQuoted, _) => TokenKind::IllegalDoubleQuoted,
-            (_, TokenKind::IllegalDoubleQuoted) => TokenKind::IllegalDoubleQuoted,
-            (_, _) => kind,
-        },
-    );
-
-    let (rest, (first, token_kind, third)) = alt((
-        (tag("'"), parse_single_quoted_value, cut(tag("'"))),
-        (tag(r#"""#), parse_double_quoted_value, tag(r#"""#)),
-    ))
-    .parse(input)?;
-
-    let parsed = Span { source: input.span.source, start: first.span.start, end: third.span.end };
-    // shorten the input text by the length of parsed
-    // this works because the original text is a slice [parsed.start..input.end] of the source
-    // so we are taking the subslice [(parsed.start+0)..parsed.end] of the source
-    let parsed =
-        SpanView { span: parsed, text: &input.text[..(parsed.end - parsed.start) as usize] };
-    Ok((rest, Token { span: parsed, kind: token_kind }))
-}
-
-#[derive(Debug)]
-enum ParseTokenError {
-    Error(String),
-    TruncatedInput(nom::Needed),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TokenKind {
-    /// field name, directly or single or double quoted
-    Value,
-    /// Single quoted value containing illegal characters
-    IllegalSingleQuoted,
-    /// Double quoted value containing illegal characters
-    IllegalDoubleQuoted,
-    /// Illegal character
-    IllegalCharacter,
-    /// number
-    FloatValue,
-    /// (
-    LeftParens,
-    /// )
-    RightParens,
-    /// [
-    LeftSquareBracket,
-    /// ]
-    RightSquareBracket,
-    /// OR
-    Or,
-    /// AND
-    And,
-    /// NOT
-    Not,
-    /// IN
-    In,
-    /// EXISTS
-    Exists,
-    /// IS
-    Is,
-    /// NULL
-    Null,
-    /// TO
-    To,
-    /// _geoRadius
-    GeoRadius,
-    /// _geoBoundingBox
-    GeoBoundingBox,
-    /// _geoPolygon
-    GeoPolygon,
-    /// _vectors
-    Vectors,
-    /// _foreign
-    Foreign,
-    /// ,
-    Comma,
-    /// =
-    Equal,
-    /// !=
-    Different,
-    /// >
-    GreaterThan,
-    /// >=
-    GreaterOrEqual,
-    /// <
-    LowerThan,
-    /// <=
-    LowerOrEqual,
-    /// End of input
-    Eof,
-}
-
-impl TokenKind {}
-
 struct ParsingContext<'a> {
     polarity: bool,
     state: ParsingState<'a>,
@@ -749,173 +403,6 @@ struct Terminal {
     kind: TerminalKind,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-/// A view of a span that contains the text of that span.
-pub struct SpanView<'a> {
-    span: Span,
-    text: &'a str,
-}
-
-impl<'a> SpanView<'a> {
-    /// Build a view from a source and its handle, spanning the entire source
-    fn from_entire_source(source_handle: SourceHandle, source: &'a FilterSource) -> Self {
-        let span = Span::from_entire_source(source_handle, source);
-        Self { span, text: &source.source }
-    }
-
-    /// A copy of the underlying span.
-    pub fn span(&self) -> Span {
-        self.span
-    }
-
-    /// build a span view from a span and the sources.
-    ///
-    /// The sources and the span must match
-    pub fn from_sources(span: Span, sources: &'a FilterSources) -> Result<Self, SpanViewError> {
-        let Some(source): Option<&FilterSource> = sources.sources.get(span.source as usize) else {
-            return Err(SpanViewError::UnknownSourceHandle(span));
-        };
-
-        let Some(text) = source.source.get(span.start as usize..span.end as usize) else {
-            return Err(SpanViewError::SpanOutOfRange(span));
-        };
-
-        Ok(Self { span, text })
-    }
-
-    /// If the text of this span is quoted in single or double quotes, then returns
-    /// the inner portion, unescaping any contained `\'`, `\"` and `\\`.
-    ///
-    /// If any unescaping occurs, then the returned Cow will be owned. Otherwise, it will be borrowed.
-    ///
-    /// If the text of this span is not quoted, returns the entire text of the span as borrowed, without unescaping anything.
-    pub fn unquote(&self) -> Cow<'a, str> {
-        if self.text.len() < 2 {
-            return Cow::Borrowed(self.text);
-        }
-        if self.text.starts_with('\'') && self.text.ends_with('\'') {
-            let inner = &self.text[1..(self.text.len() - 1)];
-            if !inner.contains('\\') {
-                return Cow::Borrowed(inner);
-            } else {
-                let inner: String = serde_json::from_reader(std::io::Read::chain(
-                    std::io::Read::chain("\"".as_bytes(), inner.as_bytes()),
-                    "\"".as_bytes(),
-                ))
-                .unwrap();
-                Cow::Owned(inner)
-            }
-        } else if self.text.starts_with('"') && self.text.ends_with('"') {
-            let inner: Cow<'_, str> = serde_json::from_str(self.text).unwrap();
-            inner
-        } else {
-            Cow::Borrowed(self.text)
-        }
-    }
-
-    /// A reference to the text corresponding to this span in the underlying source.
-    ///
-    /// As the span might be quoted and contain escaped characters, the text returned by
-    /// this method is not suitable for e.g. looking a field in a fields ids map.
-    ///
-    /// For these uses, the text returned by [`Self::unquote`] is suitable, but may incur an allocation.
-    fn raw_possibly_escaped_text(&self) -> &'a str {
-        self.text
-    }
-}
-
-// nom trait implementations
-
-impl<'a> Input for SpanView<'a> {
-    type Item = <&'a str as Input>::Item;
-
-    type Iter = <&'a str as Input>::Iter;
-
-    type IterIndices = <&'a str as Input>::IterIndices;
-
-    fn input_len(&self) -> usize {
-        self.text.len()
-    }
-
-    fn take(&self, index: usize) -> Self {
-        let split: u32 = self.span.start + u32::try_from(index).unwrap_wip();
-        Self {
-            text: Input::take(&self.text, index),
-            span: Span { source: self.span.source, start: self.span.start, end: split },
-        }
-    }
-
-    fn take_from(&self, index: usize) -> Self {
-        let split: u32 = self.span.start + u32::try_from(index).unwrap_wip();
-        Self {
-            text: Input::take_from(&self.text, index),
-            span: Span { source: self.span.source, start: split, end: self.span.end },
-        }
-    }
-
-    fn take_split(&self, index: usize) -> (Self, Self) {
-        let split: u32 = self.span.start + u32::try_from(index).unwrap_wip();
-        let (after, before) = Input::take_split(&self.text, index);
-        (
-            Self {
-                text: after,
-                span: Span { source: self.span.source, start: split, end: self.span.end },
-            },
-            Self {
-                text: before,
-                span: Span { source: self.span.source, start: self.span.start, end: split },
-            },
-        )
-    }
-
-    fn position<P>(&self, predicate: P) -> Option<usize>
-    where
-        P: Fn(Self::Item) -> bool,
-    {
-        Input::position(&self.text, predicate)
-    }
-
-    fn iter_elements(&self) -> Self::Iter {
-        Input::iter_elements(&self.text)
-    }
-
-    fn iter_indices(&self) -> Self::IterIndices {
-        Input::iter_indices(&self.text)
-    }
-
-    fn slice_index(&self, count: usize) -> Result<usize, nom::Needed> {
-        Input::slice_index(&self.text, count)
-    }
-}
-
-impl<'a, 'b> nom::Compare<&'a str> for SpanView<'b> {
-    fn compare(&self, t: &'a str) -> nom::CompareResult {
-        nom::Compare::compare(&self.text, t)
-    }
-
-    fn compare_no_case(&self, t: &'a str) -> nom::CompareResult {
-        nom::Compare::compare_no_case(&self.text, t)
-    }
-}
-
-impl<'a> nom::Offset for SpanView<'a> {
-    fn offset(&self, second: &Self) -> usize {
-        let second = second.span.start as usize;
-        let first = self.span.start as usize;
-        second - first
-    }
-}
-
-/// Error when trying to obtain a view inside a source.
-///
-/// Should typically be an internal error, as it happens if the span doesn't match the source,
-/// which is a code error.
-#[derive(Debug, Clone, Copy)]
-pub enum SpanViewError {
-    UnknownSourceHandle(Span),
-    SpanOutOfRange(Span),
-}
-
 /// A terminal typically represents the leaf objects of a filter
 ///
 /// In Meilisearch's case, it generally resolves to roaring bitmaps representing lists of docids.
@@ -974,45 +461,5 @@ impl<'a> VectorFilterView<'a> {
 }
 
 fn wip() {
-    wip::fixme!("foreign should be a terminal with id, and we should tuck the parsed instructions in a special object of the filter");
     wip::fixme!("source label should be a type rather than a string. perhaps a trait so that it can be injected from Meilisearch");
-}
-
-#[cfg(test)]
-mod test {
-    use crate::{FilterSource, Link, SpanView, Token};
-
-    #[test]
-    fn list_tokens() {
-        let source = FilterSource {
-            label: "".into(),
-            source:
-                r#"(doggo.name = Intel AND doggo.age > 12) OR "with spaces" != "some \"  \\ value" "#
-                    .into(),
-            previous_link: Link::And,
-            next_down: 0,
-        };
-        let view = SpanView::from_entire_source(0, &source);
-
-        for token in Token::iter_tokens(view) {
-            let token = token.unwrap();
-            dbg!(token);
-        }
-    }
-
-    #[test]
-    fn fail_tokens() {
-        let source = FilterSource {
-            label: "".into(),
-            source: "illegal {}".into(),
-            previous_link: Link::And,
-            next_down: 0,
-        };
-        let view = SpanView::from_entire_source(0, &source);
-
-        for token in Token::iter_tokens(view).take(10) {
-            dbg!(&token);
-            let token = token.unwrap();
-        }
-    }
 }
